optimized

+++ haplomap/src/pca.cpp
@@ -143,8 +143,8 @@ int main_eigen(int argc, char **argv) {
     std::ofstream output;
     output.open(opts->outputFileName);
     if (output.is_open()) {
-        for (int i = 0; i < results->size1; ++i) {
-            for (int j = 0; j < results->size2; ++j)
+        for (size_t i = 0; i < results->size1; ++i) {
+            for (size_t j = 0; j < results->size2; ++j)
                 output << gsl_matrix_get(results, i, j) << "\t";
             output << std::endl;
         }
